+++ instrument-service/src/main/java/sum25/group03/instrumentservice/repository/InstrumentRepository.java
@@ -9,7 +9,7 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface InstrumentRepository extends JpaRepository<Instrument, Integer> {
+public interface InstrumentRepository extends JpaRepository<Instrument, Long> {
     Optional<Instrument> findByInstrumentCode(String instrumentCode);
     List<Instrument> findByStatus(InstrumentStatus status);
 }
